Refactor main
Instead of `Text`, templates are now of type `Template`; they have a
filename and contents, both of type `Text`.

The options parser is now using `Options.Applicative` directly; this
allows for default arguments (wanted for the directory in which to place
the component).

`copyTemplateDir` was refactored to pull out the component
generation (`generateComponent`) and reduce repetition.