Add test for component directory generation
Adds monadic quickcheck test for component directory creation.

Rewrites Main to use lenses for getting fields from settings; this is to
facilitate setting the component directory with a `/tmp/` prefix so that
tests are written there.

TODO: Refactor test. Add testing for whether container/native files are
generated when the settings are appropriate.