Improve config file generation & parse failure handling
Config file is now generated in the directory that `generate-component
init` is called from; no longer attempts to find a project's root
directory. This puts onus on the user to put the config file in a sane
location.

If there's a parse failure for the config file, a default config is
used. This is still replaced with command line arguments if they are
provided.