Update the init command to take arguments
init can now take a project type, component type, and default directory.

Logic for merging is that command line takes precedent over the default
values set in defaultConfig.

Had to update the default directory to look for the existence of an
'app/components' directory and use that if present, else the current
dir. This directory structure may be unique to my workflow; might be
worth updating the default, but it is also trivial to specify your own
default.